+++ CHANGELOG.md
@@ -8,12 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support the new `.dsdl` extension for cyphal data types [#12](https://github.com/samcrow/canadensis/pull/12)
+- `canadensis_dsdl_frontend`: Support the new `.dsdl` extension for cyphal data types [#12](https://github.com/samcrow/canadensis/pull/12)
 
 ### Fixed
 
-- `register_client` example now compiles correctly after an update to `heapless` added
+- `canadensis`: `register_client` example now compiles correctly after an update to `heapless` added
   a new restriction on the size of some containers
+
+### Changed
+
+- `canadensis_dsdl_frontend`: The public regulated data types in `canadensis_dsdl_frontend` are now a submodule. This should not
+  impact any downstream code, but `git clone --recursive` is now required to make all the tests pass.
+  [#13](https://github.com/samcrow/canadensis/pull/13)
 
 ## [canadensis_dsdl_frontend-v0.3.1](https://github.com/samcrow/canadensis/tree/canadensis_dsdl_frontend-v0.3.1) - 2022-06-10